# perf: remove `async` from dispatch; will be 16% faster